Tests/LibWeb/Text/input/DOM/DOMParser-fires-details-ontoggle.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(done => {
        new DOMParser().parseFromString("<details open>", "text/html").querySelector("details").ontoggle = function(e) {
            println(`DOMParser toggle event is ${e.target.open}`);
            done();
        };
    });
</script>
