Tests/LibWeb/Text/expected/DOM/DOMParser-fires-details-ontoggle.txt

DOMParser toggle event is true
